requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pydantic-settings 
PyJWT
